# backend/tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AuthorProfileResponse": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "books": {
            "items": {
              "$ref": "#/components/schemas/BookSummaryResponse"
            },
            "title": "Books",
            "type": "array"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "bio",
          "books"
        ],
        "title": "AuthorProfileResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SchedulePublishRequest": {
        "properties": {
          "publishAt": {
            "format": "date-time",
            "title": "Publishat",
            "type": "string"
          }
        },
        "required": [
          "publishAt"
        ],
        "title": "SchedulePublishRequest",
        "type": "object"
      },
...
      "SetDiscountRequest": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "until"
        ],
        "title": "SetDiscountRequest",
        "type": "object"
      },
      "SetIsbnRequest": {
        "properties": {
          "isbn": {
            "title": "Isbn",
            "type": "string"
          }
        },
        "required": [
          "isbn"
        ],
        "title": "SetIsbnRequest",
        "type": "object"
      },
...
      "SetPriceRequest": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          }
        },
        "required": [
          "amount"
        ],
        "title": "SetPriceRequest",
        "type": "object"
      },
...
      "StoreListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "StoreListResponse",
        "type": "object"
      },
...
      "UpdateMetaRequest": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          }
        },
        "title": "UpdateMetaRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "\ud55c\uc904 ebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/authors/{author_id}": {
      "get": {
        "operationId": "author_profile_api_authors__author_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "author_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Author Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Author Profile",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}": {
      "delete": {
        "description": "\ucc45 \uc0ad\uc81c \u2014 \uc18c\uc720 \uc791\uac00\ub9cc. \ud310\ub9e4(\uc8fc\ubb38) \uc774\ub825 \uc788\uc73c\uba74 409(\ucd9c\ud310 \ucde8\uc18c\ub9cc \uac00\ub2a5).",
        "operationId": "delete_book_api_books__book_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Book",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}/discount": {
      "put": {
        "description": "\uae30\uac04 \ud560\uc778 \uc124\uc815 (\uc885\ub8cc\uc2dc\uac01\uae4c\uc9c0 \ud560\uc778\uac00 \uc801\uc6a9).",
        "operationId": "set_discount_api_books__book_id__discount_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDiscountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Discount",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}/isbn": {
      "put": {
        "operationId": "set_isbn_api_books__book_id__isbn_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetIsbnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Isbn",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/books/{book_id}/meta": {
      "put": {
        "description": "\ubd80\uc81c\u00b7\uc18c\uac1c\u00b7\ubd84\ub958 \ud3b8\uc9d1 (\uc2a4\ud1a0\uc5b4 \ub178\ucd9c\u00b7\uac80\uc0c9 \ud488\uc9c8).",
        "operationId": "update_meta_api_books__book_id__meta_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Meta",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/books/{book_id}/onix": {
      "get": {
        "description": "\ucc45 \uba54\ud0c0 \u2192 ONIX 3.0 XML (\uc11c\uc810 \uc720\ud1b5 \ud45c\uc900 \ud53c\ub4dc).",
        "operationId": "book_onix_api_books__book_id__onix_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Book Onix",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}/price": {
      "put": {
        "operationId": "set_price_api_books__book_id__price_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Price",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/books/{book_id}/publish": {
      "post": {
        "operationId": "publish_api_books__book_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/books/{book_id}/publish-now": {
      "post": {
        "description": "\uc989\uc2dc \ucd9c\uac04 (\uc2ec\uc0ac \uc0dd\ub7b5). \uc774\ubbf8 \ucd9c\ud310\ub41c \ucc45 \uc7ac\ubc1c\ud589 = \uac1c\uc815\ud310 \u2192 \uad6c\ub9e4\uc790 \uc54c\ub9bc.",
        "operationId": "publish_now_api_books__book_id__publish_now_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Now",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}/schedule": {
      "post": {
        "description": "\uc608\uc57d \ubc1c\ud589 \u2014 \uc9c0\uc815 \uc2dc\uac01\uc5d0 \uc790\ub3d9 \uac8c\uc2dc.",
        "operationId": "schedule_publish_api_books__book_id__schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulePublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Schedule Publish",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/books/{book_id}/submit": {
      "post": {
        "operationId": "submit_for_review_api_books__book_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit For Review",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/books/{book_id}/unpublish": {
      "post": {
        "description": "\ucd9c\ud310 \ucde8\uc18c \u2014 \uc2a4\ud1a0\uc5b4\uc5d0\uc11c \ube44\uacf5\uac1c\ub85c \ub0b4\ub9bc.",
        "operationId": "unpublish_api_books__book_id__unpublish_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unpublish",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/me/books": {
      "get": {
        "operationId": "my_books_api_me_books_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Books",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/store/books": {
      "get": {
        "operationId": "store_list_api_store_books_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Store List",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/store/books/{book_id}": {
      "get": {
        "operationId": "store_detail_api_store_books__book_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Detail",
        "tags": [
          "catalog"
        ]
      }
    }
  }
}